+++ xrpl/transaction/escrow_finish.go
@@ -30,6 +30,10 @@ var (
 // ````
 type EscrowFinish struct {
 	BaseTx
+	// Set of Credentials to authorize a deposit made by this transaction.
+	// Each member of the array must be the ledger entry ID of a Credential entry in the ledger.
+	// For details see https://xrpl.org/docs/references/protocol/transactions/types/payment#credential-ids
+	CredentialIDs types.CredentialIDs `json:",omitempty"`
 	// Address of the source account that funded the held payment.
 	Owner types.Address
 	// Transaction sequence of EscrowCreate transaction that created the held payment to finish.
@@ -67,6 +71,10 @@ func (e *EscrowFinish) Flatten() FlatTransaction {
 		flattened["Fulfillment"] = e.Fulfillment
 	}
 
+	if len(e.CredentialIDs) > 0 {
+		flattened["CredentialIDs"] = e.CredentialIDs.Flatten()
+	}
+
 	return flattened
 }
 
@@ -83,6 +91,10 @@ func (e *EscrowFinish) Validate() (bool, error) {
 
 	if e.OfferSequence == 0 {
 		return false, ErrEscrowFinishMissingOfferSequence
+	}
+
+	if e.CredentialIDs != nil && !e.CredentialIDs.IsValid() {
+		return false, ErrInvalidCredentialIDs
 	}
 
 	return true, nil